requirements.txt
# 2.3 is the last version that supports python 3.4-3.5
# connexion requires werkzeug but connexion < 2.4.0 does not install werkzeug
# we must peg werkzeug versions below to fix connexion
# https://github.com/zalando/connexion/pull/1044
setuptools
Flask
python-dotenv
py
pip
attrs
wheel
jinja2
tox
filelock
pluggy
six
tomli
packaging
pytest
pyyaml
markupsafe
werkzeug
click
itsdangerous
iniconfig
certifi
coverage
python-dateutil
idna
urllib3
connexion
clickclick
jsonschema
requests
inflection
pyparsing
pyrsistent
virtualenv
distlib
platformdirs
zlib
openssl
python
bcrypt
werkzeug
connexion
